requirements.txt
#openai tools
openai==1.7.0

numpy==1.26.4
tqdm
ipykernel
matplotlib
pandas

# black
# isort
# pre-commit

# parsing links
newspaper3k
lxml_html_clean

# langchain tools
langchain
langchain-openai
langgraph
pydantic
pydantic-settings

# crawl4ai
# Crawl4AI

# env setup tools
python-dotenv
